+++ zlt-commons/zlt-ribbon-spring-boot-starter/src/main/java/com/central/common/ribbon/annotation/EnableBaseFeignInterceptor.java
@@ -1,6 +1,5 @@
 package com.central.common.ribbon.annotation;
 
-import com.central.common.ribbon.config.FeignHttpInterceptorConfig;
 import com.central.common.ribbon.config.FeignInterceptorConfig;
 import org.springframework.context.annotation.Import;
 
@@ -10,13 +9,17 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * 开启feign拦截器传递数据给下游服务，包含基础数据和http的相关数据
+ * 开启feign拦截器传递数据给下游服务，只包含基础数据
  *
  * @author zlt
+ * @date 2019/10/26
+ * <p>
+ * Blog: https://blog.csdn.net/zlt2000
+ * Github: https://github.com/zlt2000
  */
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@Import({FeignInterceptorConfig.class, FeignHttpInterceptorConfig.class})
-public @interface EnableFeignInterceptor {
+@Import(FeignInterceptorConfig.class)
+public @interface EnableBaseFeignInterceptor {
 
 }
