+++ zlt-gateway/sc-gateway/src/main/java/com/central/SCGatewayApp.java
@@ -1,5 +1,6 @@
 package com.central;
 
+import com.central.common.ribbon.annotation.EnableBaseFeignInterceptor;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
@@ -13,6 +14,7 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
  * Github: https://github.com/zlt2000
  */
 @EnableFeignClients
+@EnableBaseFeignInterceptor
 @EnableDiscoveryClient
 @SpringBootApplication
 public class SCGatewayApp {
